% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agencies.R
\docType{data}
\name{agencies}
\alias{agencies}
\title{St. Louis City and County Police Departments}
\format{A data frame with 79 rows and 5 variables:
\describe{
  \item{id}{agency id number}
  \item{name}{agency name}
  \item{valid}{did agency make any traffic stops in 2017?}
  \item{countyfp}{county FIPS code}
  \item{namelsad}{county name}
  }}
\source{
Missouri Attorney General's Office
}
\usage{
data(agencies)
}
\description{
A tibble containing agency names, id numbers,
    and the county that they serve.
}
\examples{
str(agencies)
head(agencies)

}
\keyword{datasets}
