% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_agency.R
\name{mv_agency}
\alias{mv_agency}
\title{Download Agency Data}
\usage{
mv_agency(browser, agency, statistic, pause = 3)
}
\arguments{
\item{browser}{A browser binding created with \link{mv_browse}}

\item{agency}{An agency name (currently only supports \code{"SLMPD"})}

\item{statistic}{The statistic to download}

\item{pause}{Number of seconds to pause while pages load - adjust as
needed depending on internet connection}
}
\description{
Download Agency Data
}
