% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{mv_batch_agency}
\alias{mv_batch_agency}
\title{Batch Agency Downloads}
\usage{
mv_batch_agency(browser, agency, statistic, format, category, year, pause = 3)
}
\arguments{
\item{browser}{Name of a browser binding created with \code{RSelenium}}

\item{agency}{An agency name (for St. Louis City and County agencies) or id number
(for agencies statewide).}

\item{statistic}{The statistic to download (see \code{\link{mv_get_agency}})}

\item{format}{Style to reformat raw data (see \code{\link{mv_reformat}})}

\item{category}{Category to extract (see \code{\link{mv_filter}})}

\item{year}{Year to extract (see \code{\link{mv_filter}})}

\item{pause}{Number of seconds to pause while pages load - adjust as
needed depending on internet connection}
}
\value{
A tibble with the formatted and subset data.
}
\description{
Designed to be including in a \code{purrr} call. This wraps
    \code{\link{mv_get_agency}}, \code{\link{mv_reformat}}, and \code{\link{mv_filter}}.
}
