% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_browse.R
\name{mv_browse}
\alias{mv_browse}
\title{Open Browser}
\usage{
mv_browse(browser = "chrome", chromever)
}
\arguments{
\item{browser}{Browser type to pass to \code{RSelenium::rdDrive}}

\item{chromever}{Chrome version to pass to \code{RSelenium::rdDrive}}
}
\description{
Open Browser
}
