% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shed.R
\name{mv_create_shed}
\alias{mv_create_shed}
\title{Create Ticket Shed}
\usage{
mv_create_shed(.data, primary, radius, power = 1)
}
\arguments{
\item{.data}{A sf object}

\item{primary}{\code{GEOID} value of the municipality that the ticket shed is for}

\item{radius}{Euclidean distance in miles from the primary feature's centroid
that defines the size of the ticket shed.}

\item{power}{Power value for inverse distance weighting}
}
\value{
An \code{sf} object with the tracts contained within the ticket shed as
    well as the distance in kilometers from the primary feature's centriod
    and the weight for each tract.
}
\description{
Creates an object with all census tracts whose centroids fall
    within the specified radius of the primary feature. This primary feature
    must be the first feature listed in the input data object. Euclidean distance
    is used for this calculation. Once the relevant tracts are selected, a weight
    is calculated for each using inverse distance weighting. The power argument
    should be used to control this calculation process.
}
