% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{mv_filter}
\alias{mv_filter}
\title{Subset Data}
\usage{
mv_filter(.data, category, year)
}
\arguments{
\item{.data}{A tbl created with \code{\link{mv_get_agency}}}

\item{category}{Category to extract}

\item{year}{Year to extract}
}
\value{
A tibble containing the subset data.
}
\description{
Subset data to extract a single category and time point for cross
    sectional comparisons between departments.
}
