% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{mv_get_agency}
\alias{mv_get_agency}
\title{Download Agency Data}
\usage{
mv_get_agency(browser, agency, statistic, pause = 3, add_agency = TRUE)
}
\arguments{
\item{browser}{A browser binding created with \link{mv_browse}}

\item{agency}{An agency name (currently only supports \code{"SLMPD"})}

\item{statistic}{The statistic to download}

\item{pause}{Number of seconds to pause while pages load - adjust as
needed depending on internet connection}

\item{add_agency}{Append agency name to returned data}
}
\description{
Download Agency Data
}
